[FlexNG] Row item container separation
While working on CL:3575366, I noticed that we were using
|broke_before_row| instead of |*broke_before_row| when calculating
if a flex row item has container separation. However, updating this
caused the test added in CL:3575366 to fail.

The reason the test failed was the first item no longer had container
separation after it broke before the first time, causing the
entire flex container to break before.

To fix this, update CalculateBreakAppealBefore() to no longer set
|breakable_at_start_of_container| to false for row items. This allows
the first item in the test to break into the next outer multicol
fragmentainer rather than forcing the whole container into the next
fragmentainer.

multi-line-row-flex-fragmentation-039.html was updated based on the
new behavior.

Bug: 660611
Change-Id: Ie92f33de1948ec2d542c6e446091d183bac30489